\name{currency.names}
\alias{currency.names}
\docType{data}
\title{
Full Currency Names for the Historical Exchange Rates Against the Euro}
\description{
This character vector contains the full currency names for the
40 currencies worldwide with reported exchange rates (against the Euro)
in the data frame eu. 
}
\usage{data(eu)}
\note{
This is one of the objects found in eu.rda. See the see also
section for related objects.
}
\source{
European Central Bank
http://www.ecb.europa.eu/stats/eurofxref/eurofxref-hist.xml
}
\seealso{
        \code{\link{eu}}
}
\keyword{datasets}
