\name{eu}
\alias{eu}
\docType{data}
\title{
Historical Exchange Rates Against the Euro}
\description{
This data frame contains currency exchange rates (against the Euro) for 40
currencies worldwide on dates from 1999-01-11  to 2009-10-30.
}
\usage{data(eu)}
\format{
  \describe{
    \item{\code{Date}}{a factor with levels corresponding to Dates in YYYY-MM-DD format}
    \item{\code{USD}}{Exchange Rate for the US Dollar}
    \item{\code{JPY}}{Exchange Rate for the Japanese Yen}
    \item{\code{BGN}}{Exchange Rate for the Bulgarian Lev}
    \item{\code{CYP}}{Exchange Rate for the Cypriot Pound}
    \item{\code{CZK}}{Exchange Rate for the Czech Koruna}
    \item{\code{DKK}}{Exchange Rate for the Danish Krone}
    \item{\code{EEK}}{Exchange Rate for the Estonian Kroon}
    \item{\code{GBP}}{Exchange Rate for the Pound Sterling}
    \item{\code{HUF}}{Exchange Rate for the Hungarian Forint}
    \item{\code{LTL}}{Exchange Rate for the Lithuanian Litas}
    \item{\code{LVL}}{Exchange Rate for the Latvian Lats}
    \item{\code{MTL}}{Exchange Rate for the Maltese Lira}
    \item{\code{PLN}}{Exchange Rate for the Polish Zloty}
    \item{\code{ROL}}{Exchange Rate for the Romanian Lei}
    \item{\code{RON}}{Exchange Rate for the Romanian Leu}
    \item{\code{SEK}}{Exchange Rate for the Swedish Krona}
    \item{\code{SIT}}{Exchange Rate for the Slovenian Tolar}
    \item{\code{SKK}}{Exchange Rate for the Slovak Koruna}
    \item{\code{CHF}}{Exchange Rate for the Swiss Franc}
    \item{\code{ISK}}{Exchange Rate for the Icelandic Krona}
    \item{\code{NOK}}{Exchange Rate for the Norwegian Krone}
    \item{\code{HRK}}{Exchange Rate for the Croatian Kuna}
    \item{\code{RUB}}{Exchange Rate for the Russian Rouble}
    \item{\code{TRL}}{Exchange Rate for the Turkish Lire}
    \item{\code{TRY}}{Exchange Rate for the Turkish Lira}
    \item{\code{AUD}}{Exchange Rate for the Australian Dollar}
    \item{\code{BRL}}{Exchange Rate for the Brasilian Real}
    \item{\code{CAD}}{Exchange Rate for the Canadian Dollar}
    \item{\code{CNY}}{Exchange Rate for the Chinese Yuan Renminbi}
    \item{\code{HKD}}{Exchange Rate for the Hong Kong Dollar}
    \item{\code{IDR}}{Exchange Rate for the Indonesian Rupiah}
    \item{\code{INR}}{Exchange Rate for the Indian Rupee}
    \item{\code{KRW}}{Exchange Rate for the South Korean Won}
    \item{\code{MXN}}{Exchange Rate for the Mexican Peso}
    \item{\code{MYR}}{Exchange Rate for the Malaysian Ringgit}
    \item{\code{NZD}}{Exchange Rate for the New Zealand Dollar}
    \item{\code{PHP}}{Exchange Rate for the Philippine Peso}
    \item{\code{SGD}}{Exchange Rate for the Singapore Dollar}
    \item{\code{THB}}{Exchange Rate for the Thai Baht}
    \item{\code{ZAR}}{Exchange Rate for the South African Rand}
  }
}
\source{
European Central Bank
http://www.ecb.europa.eu/stats/eurofxref/eurofxref-hist.xml
}
\keyword{datasets}
